requirements.txt
acres==0.2.0
altair==5.5.0
annotated-types==0.7.0
attrs==25.1.0
blinker==1.9.0
Brotli==1.1.0
cachetools==5.5.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
ci-info==0.3.0
click==8.1.8
configobj==5.0.9
configparser==7.1.0
cryptography==44.0.1
cssselect2==0.7.0
etelemetry==0.3.1
filelock==3.17.0
fitz==0.0.1.dev2
fonttools==4.56.0
fpdf==1.7.2
gitdb==4.0.12
GitPython==3.1.44
google-ai-generativelanguage==0.6.15
google-api-core==2.24.1
google-api-python-client==2.160.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.4
googleapis-common-protos==1.67.0rc1
greenlet==3.1.1
grpcio==1.70.0
grpcio-status==1.70.0
httplib2==0.22.0
idna==3.10
iniconfig==2.0.0
isodate==0.6.1
Jinja2==3.1.5
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
looseversion==1.3.0
lxml==5.3.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
narwhals==1.26.0
nibabel==5.3.2
nipype==1.9.2
numpy==2.2.2
packaging==24.2
pandas==2.2.3
pathlib==1.0.1
pdfkit==1.0.0
pdfminer.six==20231228
pdfplumber==0.11.5
pillow==11.1.0
playwright==1.50.0
pluggy==1.5.0
proto-plus==1.26.0
protobuf==5.29.3
prov==2.0.1
puremagic==1.28
pyarrow==19.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pydeck==0.9.1
pydot==3.0.4
pydyf==0.11.0
pyee==12.1.1
Pygments==2.19.1
PyMuPDF==1.25.3
pyparsing==3.2.1
pypdfium2==4.30.1
pyphen==0.17.2
pytesseract==0.3.13
pytest==8.3.4
pytest-base-url==2.1.0
pytest-playwright==0.7.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-slugify==8.0.4
pytz==2025.1
pyxnat==1.6.3
rdflib==6.3.2
referencing==0.36.2
requests==2.32.3
rich==13.9.4
rpds-py==0.22.3
rsa==4.9
scipy==1.15.1
simplejson==3.19.3
six==1.17.0
smmap==5.0.2
streamlit==1.42.0
tenacity==9.0.0
text-unidecode==1.3
tinycss2==1.4.0
tinyhtml5==2.0.0
toml==0.10.2
tornado==6.4.2
tqdm==4.67.1
traits==7.0.2
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0
weasyprint==64.0
webencodings==0.5.1
zopfli==0.2.3.post1
